pyproject.toml
[tool.poetry]
name = "multipepgen"
version = "1.0.0"
description = "A neural network-based model for generating antimicrobial peptide sequences using Conditional GAN"
authors = ["A. Orrego <aorregope@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
numpy = ">=1.21.0"
pandas = ">=1.3.0"
scikit-learn = ">=1.0.0"
tensorflow = ">=2.8.0"
keras = ">=2.8.0"
pyyaml = ">=6.0"
tqdm = ">=4.62.0"
matplotlib = ">=3.5.0"
seaborn = ">=0.11.0"
plotly = ">=5.0.0"
scipy = ">=1.7.0"
biopython = ">=1.79"
joblib = ">=1.1.0"
imageio = ">=2.9.0"
xgboost = ">=1.5.0"
modlamp = ">=4.3.0"

[tool.poetry.group.dev.dependencies]
pytest = ">=6.0.0"
pytest-cov = ">=3.0.0"
black = ">=22.0.0"
flake8 = ">=4.0.0"
mypy = ">=0.950"

[tool.poetry.group.docs.dependencies]
sphinx = ">=4.0.0"
sphinx-rtd-theme = ">=1.0.0"

[tool.poetry.group.notebooks.dependencies]
jupyter = ">=1.0.0"
ipykernel = ">=6.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api" 